demo/Python/requirements.txt
opencv-python
matplotlib
pyquaternion
pandas
colorama
xgboost-cpu